[!] Centered planar transform

A data transformation according to the centered planar transformation is done. Default method is function cpt from package compositions.

cpt(x, ...)

# S3 method for default
cpt(x, ...)

# S3 method for hyperSpec
cpt(x, ...)

Arguments

x
A numeric data frame or matrix.

...
further parameters to cpt.

Value

Returns the transformed data matrix with the same dimension as x.

See also

cpt

Other row-wise transformations: alr, apt, clr, ilr, ipt

Examples

cpt(Spectra2)

#> Warning: Calling 'structure(NULL, *)' is deprecated, as NULL cannot have attributes.
#>   Consider 'structure(list(), *)' instead.

#> hyperSpec object
#>    150 spectra
#>    3 data columns
#>    501 data points / spectrum
#> wavelength: lambda/nm [integer] 300 301 ... 800 
#> data:  (150 rows x 3 columns)
#>    1. gr:  [factor] B B ... A 
#>    2. class:  [factor] N l ... S1 
#>    3. spc: I / a.u. [matrix501] -0.0005372106 -0.0003909578 ... -0.001848332 

plot(Spectra2)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

plot(cpt(Spectra2))

#> Warning: Calling 'structure(NULL, *)' is deprecated, as NULL cannot have attributes.
#>   Consider 'structure(list(), *)' instead.

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'