DEPRECATED!

DEPRECATED! Create folds with blocking and stratification (BS) for (repeated) k-fold cross-validation Function randomly divides observations into folds that are used for (repeated) k-fold cross-validation. In these folds observations are:

blocked by values in variable ID (i.e. observations with the same "ID" are treated as one unit (a block) and are always in the same fold);

stratified by levels of factor variable gr (the proportions of these grouped units of observations per each group (level) are kept aproximately constant throughout all folds).

createFolds2(data = NULL, gr = NULL, ID = NULL, k = 5,
  returnTrain = TRUE, ...)

stratifiedFolds(data = NULL, gr = NULL, ID = NULL, k = 5,
  returnTrain = TRUE, ...)

createFoldsBS(data = NULL, stratify_by = NULL, block_by = NULL, k = 5,
  returnTrain = TRUE, times = 1, seeds = NULL, gr = stratify_by,
  ID = block_by)

Arguments

data
A data frame, that contains variables which names are denoted by arguments ID and by gr.

k
(integer) A number of folds, default k = 5.

returnTrain
(logical) If TRUE, returns indices of variables in training set. If FALSE, returns indices of variables in test set.

stratify_by, gr
A vector or a name of factor variable in data, which levels will be used for stratification. E.g., vector with medical groups.

block_by, ID
A vector or a name of variable in data, that contains identification codes/numbers (ID). These codes will be used for blocking.

times
(integer) number of repetitions for repeated cross-vatidtion.

seeds
(vector of integers | NULL) Seeds for random number generator for each repetition. If seeds = NULL random seeds are generated. If number of repetitions is greater than number of provided seeds, random seeds are generated and added to the provided ones. The first seed will be used to ensure reproducibility of the randomly generated seeds.

(See set.seed for more information about random number generation).

Value

A list of folds. In each fold there are indices observations. The structure of outpus is the similar to one created with createFolds().

Note

If k is such big, that some folds have no observations of a certain group (i.e. level in gr), an error is returned. In that case smaller value of k is recommended. createFolds2, stratifiedFolds is a wrapper of createFoldsBS.

See also

createFolds Test if folds are blocked and stratified cvo_test_bs

Examples

library(spHelper)

# Load data
     data("DataSet1")

# Explore data
     str(DataSet1)

#> 'data.frame':	40 obs. of  3 variables:
#>  $ ID  : Factor w/ 20 levels "1","2","3","4",..: 1 1 2 2 3 3 4 4 5 5 ...
#>  $ gr  : Factor w/ 4 levels "A","B","C","D": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ .row: int  1 2 3 4 5 6 7 8 9 10 ...

     table(DataSet1[,c("gr","ID")])

#>    ID
#> gr  1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20
#>   A 2 2 2 2 2 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0
#>   B 0 0 0 0 0 2 2 2 2  2  0  0  0  0  0  0  0  0  0  0
#>   C 0 0 0 0 0 0 0 0 0  0  2  2  2  2  2  0  0  0  0  0
#>   D 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  2  2  2  2  2

     summary(DataSet1)

#>        ID     gr          .row      
#>  1      : 2   A:10   Min.   : 1.00  
#>  2      : 2   B:10   1st Qu.:10.75  
#>  3      : 2   C:10   Median :20.50  
#>  4      : 2   D:10   Mean   :20.50  
#>  5      : 2          3rd Qu.:30.25  
#>  6      : 2          Max.   :40.00  
#>  (Other):28                         

# Explore functions
     nFolds = 5

# If variables of data frame are provided:
     Folds1_a <- createFoldsBS(data = DataSet1,
                              stratify_by = "gr", block_by = "ID",
                               k = nFolds, returnTrain = FALSE)
     # str(Folds1_a)
     cvo_test_bs(Folds1_a, DataSet1, "gr", "ID")

#> Warning: Lengths of arguments 'DATA' and 'VAR' do not match!!!

#> Error in sort.list(y): 'x' must be atomic for 'sort.list'
#> Have you called 'sort' on a list?

# If "free" variables are provided:
     Folds1_b <- createFoldsBS(gr = DataSet1$gr, ID = DataSet1$ID,
                                k = nFolds, returnTrain = FALSE)
     # str(Folds1_b)
     cvo_test_bs(Folds1_b, DataSet1, "gr", "ID")

#> Warning: Lengths of arguments 'DATA' and 'VAR' do not match!!!

#> Error in sort.list(y): 'x' must be atomic for 'sort.list'
#> Have you called 'sort' on a list?

# Not blocked but stratified
     Folds1_c <- createFoldsBS(gr = DataSet1$gr, k=nFolds, returnTrain=FALSE)
     # str(Folds1_c)
     cvo_test_bs(Folds1_c, DataSet1, "gr", "ID")

#> Warning: Lengths of arguments 'DATA' and 'VAR' do not match!!!

#> Error in sort.list(y): 'x' must be atomic for 'sort.list'
#> Have you called 'sort' on a list?

# Blocked but not stratified
     Folds1_d <- createFoldsBS(block_by = DataSet1$ID, k = nFolds, returnTrain = FALSE)
     # str(Folds1_d)
     cvo_test_bs(Folds1_d, DataSet1, "gr", "ID")

#> Warning: Lengths of arguments 'DATA' and 'VAR' do not match!!!

#> Error in sort.list(y): 'x' must be atomic for 'sort.list'
#> Have you called 'sort' on a list?