[!] Create indices of folds with blocking and stratification (cvo object)

[DESCRIBTION MUST BE UPDATED] Create indices of folds with blocking and stratification for (repeated) k-fold cross-validation. Function cvo_create_folds randomly divides observations into folds that are used for (repeated) k-fold cross-validation. In these folds observations are:

blocked by values in variable block_by (i.e. observations with the same "ID" or other kind of blocking factor are treated as one unit (a block) and are always in the same fold);

stratified by levels of factor variable stratify_by (the proportions of these grouped units of observations per each group (level) are kept approximately constant throughout all folds).

cvo_create_folds(data = NULL, stratify_by = NULL, block_by = NULL,
  k = folds, times = 1, seeds = NULL, mode = c("caret", "mlr")[1],
  returnTrain = c(TRUE, FALSE, "both")[1], predict = c("test", "train",
  "both")[1], folds = 5)

Arguments

data
A data frame, that contains variables which names are denoted by arguments block_by and by stratify_by.

stratify_by
A vector or a name of factor variable in data, which levels will be used for stratification. E.g., a vector with medical groups.

block_by
A vector or a name of variable in data, that contains identification codes/numbers (ID). These codes will be used for blocking.

k, folds
(integer) A number of folds, default k = 5.

times
(integer) a number of repetitions for repeated cross-validation.

seeds
(vector of integers | NULL) Seeds for random number generator for each repetition. If seeds = NULL random seeds are generated. If number of repetitions is greater than number of provided seeds, random seeds are generated and added to the provided ones. The first seed will be used to ensure reproducibility of the randomly generated seeds.

(See set.seed for more information about random number generation).

mode
(character) Either caret-like or mlr-like cvo object. This option is not implemented yet!

returnTrain
(logical | character) If TRUE, returns indices of variables in a training set (caret style). If FALSE, returns indices of variables in a test set (caret style). If "both", returns indices of variables in both training and test sets (mlr style).

predict
[character(1)] What to predict during resampling: “train”, “test” or “both” sets. Default is “test”.

Value

(list) A list of folds. In each fold there are indices observations. The structure of outputs is the similar to one created with either function createFolds() from caret or function makeResampleInstance() in mlr.

Note

If k is such big, that some folds have no observations of a certain group (i.e., level in stratify_by), an error is returned. In that case smaller value of k may be recommended.

See also

Function createFolds from package caret. Function makeResampleInstance from package mlr. Test if folds are blocked and stratified test_folds_BS

Other spHelper fold creation functions: cvo_get_info, cvo_test_bs

Examples

library(spHelper)

# Load data
     data("DataSet1")

# Explore data
     str(DataSet1)

#> 'data.frame':	40 obs. of  3 variables:
#>  $ ID  : Factor w/ 20 levels "1","2","3","4",..: 1 1 2 2 3 3 4 4 5 5 ...
#>  $ gr  : Factor w/ 4 levels "A","B","C","D": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ .row: int  1 2 3 4 5 6 7 8 9 10 ...

     table(DataSet1[,c("gr","ID")])

#>    ID
#> gr  1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20
#>   A 2 2 2 2 2 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0
#>   B 0 0 0 0 0 2 2 2 2  2  0  0  0  0  0  0  0  0  0  0
#>   C 0 0 0 0 0 0 0 0 0  0  2  2  2  2  2  0  0  0  0  0
#>   D 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  2  2  2  2  2

     summary(DataSet1)

#>        ID     gr          .row      
#>  1      : 2   A:10   Min.   : 1.00  
#>  2      : 2   B:10   1st Qu.:10.75  
#>  3      : 2   C:10   Median :20.50  
#>  4      : 2   D:10   Mean   :20.50  
#>  5      : 2          3rd Qu.:30.25  
#>  6      : 2          Max.   :40.00  
#>  (Other):28                         

# Explore functions
     nFolds = 5

# If variables of data frame are provided:
     Folds1_a <- cvo_create_folds(data = DataSet1,
                              stratify_by = "gr", block_by = "ID",
                               k = nFolds, returnTrain = FALSE)
     # str(Folds1_a)
     cvo_test_bs(Folds11_a, "gr", "ID", DataSet1)

#> Error in lapply(obj, make_table, By): object 'Folds11_a' not found

# If "free" variables are provided:
     Folds1_b <- cvo_create_folds(gr = DataSet1$gr, ID = DataSet1$ID,
                                k = nFolds, returnTrain = FALSE)

#> Error in cvo_create_folds(gr = DataSet1$gr, ID = DataSet1$ID, k = nFolds,     returnTrain = FALSE): unused arguments (gr = DataSet1$gr, ID = DataSet1$ID)

     # str(Folds1_b)
     cvo_test_bs(Folds11_b, "gr", "ID", DataSet1)

#> Error in lapply(obj, make_table, By): object 'Folds11_b' not found

# Not blocked but stratified
     Folds1_c <- cvo_create_folds(gr = DataSet1$gr, k=nFolds, returnTrain=FALSE)

#> Error in cvo_create_folds(gr = DataSet1$gr, k = nFolds, returnTrain = FALSE): unused argument (gr = DataSet1$gr)

     # str(Folds1_c)
     cvo_test_bs(Folds11_c, "gr", "ID", DataSet1)

#> Error in lapply(obj, make_table, By): object 'Folds11_c' not found

# Blocked but not stratified
     Folds1_d <- cvo_create_folds(block_by = DataSet1$ID, k = nFolds, returnTrain = FALSE)
     # str(Folds1_d)
     cvo_test_bs(Folds11_d, "gr", "ID", DataSet1)

#> Error in lapply(obj, make_table, By): object 'Folds11_d' not found