[!.] Wrapper function for alternating least squares multivariate curve resolution (MCR-ALS)

This function is the same as doALS from package alsace (version 1.6.0), just with more possible input parameters, which are passed to als from package ALS.

doALS2(Xl, PureS, maxiter = 100, normS = 0.5, uniC = FALSE,
  uniS = FALSE, nonnegS = TRUE, nonnegC = TRUE, optS1st = FALSE,
  baseline = FALSE, closureC = list(), ...)

Arguments

Xl
a list of (numerical) data matrices. Missing values are not allowed.

PureS
Initial estimates of pure spectral components.

maxiter
The maximum number of iterations to perform (where an iteration is optimization of either AList and C)

normS
numeric indicating whether the spectra are normalized; if normS>0, the spectra are normalized. If normS==1 the maximum of the spectrum of each component is constrained to be equal to one; if normS > 0 && normS!=1 then the norm of the spectrum of each component is constrained to be equal to one.

uniC
logical indicating whether unimodality constraints should be applied to the columns of C

uniS
logical indicating whether unimodality constraints should be applied to the columns of S

nonnegS
logical indicating whether the components (columns) of the matrix S should be constrained to non-negative values

nonnegC
logical indicating whether the components (columns) of the matrix C should be constrained to non-negative values

optS1st
logical indicating whether the first constrained least squares regression should estimate S or CList.

baseline
logical indicating whether a baseline component is present; if baseline=TRUE then this component is exempt from constraints unimodality or non-negativity

closureC
list; if the length is zero, then no closure constraints are applied. If the length is not zero, it should be equal to the number of datasets in the analysis, and contain numeric vectors consisting of the desired value of the sum of each row of the concentration matrix.

...
Other parametars to be passed to function als.

Value

The same as in doALS.

See also

Examples

# /NO examples YET/