[!] Add a variable to hyperSpec object

Add (or overwrite, if already exists) a variable which name is passed as var.name with values in unique.values that correspond to levels of factor variable by.

hyAdd(sp, var.name, by, unique.values)

Arguments

sp
hyperSpec object.

var.name
The name of variable to be added.

by
A factor variale which levels will correspond to values in unique.values.

unique.values
Unique values used to create the variavle.

Value

HyperSpec object with added/replaced variable. Lables of the variable indicate unique values used.

See also