[+] Add labels to "PAP_PD_2014" and transform the dataset

Function is designed to label data collected during investigation called "PAP_PD_2014".

hyAdd_Labels_PAP_PD_2014(sp, language = "EN")

Arguments

sp
A hyperSpec object of PAP_PD_2014 data, created by function read3csv2hy.

language
A string, indicating a language of labels. Possible entries are EN - English and LT - Lithuanian. Default is EN.

Value

A labeled object with reduced number of data columns.

Details

Select data columns, that are not removed, and add labels to hyperSpec object of "PAP_PD_2014" data.

See also

Other spHelper functions for spectroscopy and hyperSpec: IQR_outliers, binning, file, gapDer, hy2mat, hyAdd_Label_wl, hyAdd_Labels_TD2009, hyAdd, hyDrop_NA, hyGet_palette, hyRm_palette, mad_outliers, mean_Nsd, median_Nmad, plot_hyPalette, prepare_PAP_RK_2014__MATLAB_failui, read.OOIBase32, read.OceanView.header, read.OceanView, read.sp.csv2, read3csv2hy, replace_spc, sd_outliers, spStat, sp_class_perform