Return transposed matrix of spectra in hyperSpec object

Transpose matrix of spectra in hyperSpec object, add column with wavelengths and return as a dataframe

hy_spc2df(obj)

Arguments

obj
hyperSpec object

Value

A data frame with wavelengths in the first column and each row of hyperSpec object in a separate column.

Examples

  library(spHelper)

  chondro[[]][1,1:6]

#>         602         606         610         614         618         622 
#> 0.037818568 0.040815282 0.035726519 0.019315397 0.005847812 0.015234723 

  hy_spc2df(chondro)[1:6,1:2]

#>      wl          X1
#> 602 602 0.037818568
#> 606 606 0.040815282
#> 610 610 0.035726519
#> 614 614 0.019315397
#> 618 618 0.005847812
#> 622 622 0.015234723