[!] Plot distributions of spectroscopic data

[!] Plot distributions of spectroscopic data

qplot_spDistrib(sp, by = stop("`by` is missing"), percent = NULL,
  probs = NULL, palette = hyGet_palette(sp))

Arguments

sp
hyperSpec object.

by
A vector (factor variable) with indicated groups for each case: either variable name inside the object sp, or a vector of length nrow(sp).

percent
Numeric value between 0 and 100 that indicates percentage of samples nearest to median be plotted as a ribbon.

probs
verctor of size 2 with values between 0 and 1 that indicates probabilities at which percentiles should be plotted.

Value

A ggplot object. More details in package ggplot2.