[!] Read data from OceanOptics OOI Base32 ASCII file with header lines

Read contents of spectroscopic ASCII data file with header lines that contain metadata to hyperSpec object. THe file is created by software `OOI Base32 2.0.0.5`.

read.OOIBase32(file, dec = ".", n = 19,
  last_headerline_text = ">>>>>Begin Spectral Data<<<<<",
  version_ = "2.0.0.5", parse_filename = NULL, ignore.case = TRUE,
  xlab = NULL, ylab = "I, a.u.")

read.OOIBase32_2(file, dec = ",", n = 19,
  last_headerline_text = ">>>>>Begin Spectral Data<<<<<",
  version_ = "2.0.0.5", parse_filename = NULL, ...)

read.OOIBase32_0(file, dec = ".", n = 19,
  last_headerline_text = ">>>>>Begin Spectral Data<<<<<",
  version_ = "2.0.0.5", parse_filename = parser_TD2015, ...)

Arguments

file
The name of the file which the data are to be read from.

dec
("." | ",") The character used in the file for decimal points , e.g. period (".") or comma (",").

n
(integer) Number of header lines to be scanned . Default is 17. These lines include the header line indicated in last_headerline_text.

last_headerline_text
(string) A string, that indicates the last header line. Default is ">>>>>Begin Spectral Data<<<<<".

version_
version of OOI Base32 file. Curently version "2.0.0.5" is supported.

parse_filename
(function | string | NULL) Either a function that parses string of filename and extracts relevant information, or a regular expression to be used to parse file name and extract relevant information, or NULL.

ignore.case
A logical that indicates if regular extression in parse_filename is case sensitive. Applies ONLY if parse_filename is a string with regular expression.

xlab
Label for x (wavelength) axis. If NULL (defaut) - the label is automatically selected as "Wavelength, nm".

ylab
(string) Label for y (intensity) axis. Defaut is "I, a.u.".

Value

A hyperSpec object with technical and spectroscopic information from file file.

Details

In read.OOIBase32 the default decimal symbol is dot (.). In read.OOIBase32_2 - comma (,).

See also

Other spHelper functions for spectroscopy and hyperSpec: IQR_outliers, binning, file, gapDer, hy2mat, hyAdd_Label_wl, hyAdd_Labels_PAP_PD_2014, hyAdd_Labels_TD2009, hyAdd, hyDrop_NA, hyGet_palette, hyRm_palette, mad_outliers, mean_Nsd, median_Nmad, plot_hyPalette, prepare_PAP_RK_2014__MATLAB_failui, read.OceanView.header, read.OceanView, read.sp.csv2, read3csv2hy, replace_spc, sd_outliers, spStat, sp_class_perform

Other functions to read spectroscopic data: file, read.OceanView.header, read.OceanView

Examples

read.OOIBase32("Spectra_ascii.txt")

#> Warning: cannot open file 'Spectra_ascii.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

# Read several files to one `hyperspec` object:

Files   <- dir()[1:4]                           # 4 files are selected
sp_list <- lapply(Files, read.OOIBase32)  # Make a list of objects

#> Warning: No header lines were read. Value of parameter `n` may be incorrect

#> Error in if (nrows < 0L) 5 else min(5L, (header + nrows)): argument is of length zero

sp      <- collapse(sp_list)                    # Merge several spectra to one object

#> Error in collapse(sp_list): object 'sp_list' not found

plotmat(sp)

#> Error in is(object, "hyperSpec"): object 'sp' not found