[!v0.1] Read spectroscopic CSV (comma separated values) file

[!v0.1] Read spectroscopic CSV (comma separated values) file

read.sp.csv2(file, parse_filename = NULL, ylab = "Absorbance",
  xlab = "Wavelength, nm", ...)

Arguments

file
The name of the file with specroscopic data.

parse_filename
(function | string | NULL) Either a function that parses string of filename and extracts relevant information, or a regular expression to be used to parse file name and extract relevant information, or NULL.

ylab
(string) Label for y (intensity) axis. Defaut is "I, a.u.".

xlab
(string | NULL) Label for x (wavelength) axis. If NULL (defaut) - the label is selected automatically.

...
parameters to be passed to read.csv.

Value

A hyperSpec object with technical and spectroscopic information from file file.

Details

In read.sp.csv the default decimal symbol is dot (.). In read.sp.csv2 - comma (,).

See also

Other spHelper functions for spectroscopy and hyperSpec: IQR_outliers, binning, file, gapDer, hy2mat, hyAdd_Label_wl, hyAdd_Labels_PAP_PD_2014, hyAdd_Labels_TD2009, hyAdd, hyDrop_NA, hyGet_palette, hyRm_palette, mad_outliers, mean_Nsd, median_Nmad, plot_hyPalette, prepare_PAP_RK_2014__MATLAB_failui, read.OOIBase32, read.OceanView.header, read.OceanView, read3csv2hy, replace_spc, sd_outliers, spStat, sp_class_perform

Examples

read.sp.csv2("Spectra_ascii.txt")

#> Warning: cannot open file 'Spectra_ascii.txt': No such file or directory

#> Error in file(file, "rt"): cannot open the connection

# Read several files to one `hyperspec` object:

Files   <- dir()[1:4]                           # 4 files are selected
sp_list <- lapply(Files, read.sp.csv)  # Make a list of objects

#> Error in match.fun(FUN): object 'read.sp.csv' not found

sp      <- collapse(sp_list)                    # Merge several spectra to one object

#> Error in collapse(sp_list): object 'sp_list' not found

plotmat(sp)

#> Error in is(object, "hyperSpec"): object 'sp' not found