[+] Reconstruct spectra from loadings and scores (i.e. components and amplitudes)

Reconstruct spectra from loadings and scores (i.e. from spectroscopic components and their amplitudes).

reconstructSp(loadings, scores, sp = NULL)

Arguments

loadings
Components' spectra (loadings) as either a matrix or a hyperSpec object.

scores
Components' amplitudes (scores) as either a matrix or a hyperSpec object.

sp
hyperSpec object which has the same number or rows as scores. If sp is provided, reconstructed spectra will replace other spectroscopic data in sp. This object will be returned as a result of this function.

Value

Reconstruction of the original spectroscopic data matrix, based on scores and loadings as either a matrix or hyperSpec object. reconstructed <- scores %*% loadings

See also

reconstruct in ChemometricsWithR, project in ChemometricsWithR, reconstruct in Rssa

Other component analysis / factorisation related functions in spHelper: getScores, infoDim, plot_spDiff, qplot_infoDim, qplot_kAmp, qplot_kSp, qplot_spc, sortLoadings, unipeak, whichOutlier

Examples

reconstructSp(Loadings2, Scores2)

#>            [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>   [1,] 160.1654 161.0204 161.8526 162.6618 163.4481 164.2114 164.9519 165.6696
#>            [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#>   [1,] 166.3648 167.0376 167.6882 168.3171 168.9245 169.5108 170.0765 170.6221
#>           [,17]    [,18]    [,19]    [,20]    [,21]    [,22]    [,23]    [,24]
#>   [1,] 171.1482 171.6554 172.1443 172.6157 173.0704 173.5091 173.9328 174.3423
#>           [,25]    [,26]    [,27]    [,28]    [,29]    [,30]    [,31]    [,32]
#>   [1,] 174.7388 175.1232 175.4966 175.8602 176.2152 176.5629 176.9044 177.2413
#>           [,33]    [,34]    [,35]    [,36]    [,37]    [,38]    [,39]    [,40]
#>   [1,] 177.5749 177.9066 178.2380 178.5705 178.9058 179.2454 179.5910 179.9443
#>           [,41]    [,42]    [,43]    [,44]    [,45]    [,46]    [,47]    [,48]
#>   [1,] 180.3069 180.6805 181.0671 181.4681 181.8856 182.3212 182.7767 183.2538
#>           [,49]    [,50]    [,51]    [,52]    [,53]    [,54]    [,55]    [,56]
#>   [1,] 183.7545 184.2803 184.8331 185.4146 186.0265 186.6703 187.3476 188.0602
#>           [,57]    [,58]    [,59]    [,60]    [,61]    [,62]    [,63]    [,64]
#>   [1,] 188.8093 189.5964 190.4230 191.2901 192.1991 193.1510 194.1468 195.1874
#>           [,65]    [,66]    [,67]    [,68]    [,69]    [,70]    [,71]    [,72]
#>   [1,] 196.2735 197.4058 198.5848 199.8109 201.0843 202.4052 203.7734 205.1889
#>           [,73]    [,74]    [,75]    [,76]    [,77]    [,78]    [,79]    [,80]
#>   [1,] 206.6512 208.1598 209.7141 211.3132 212.9562 214.6418 216.3688 218.1357
#>           [,81]    [,82]    [,83]    [,84]    [,85]    [,86]    [,87]    [,88]
#>   [1,] 219.9408 221.7823 223.6583 225.5667 227.5052 229.4715 231.4631 233.4774
#>           [,89]    [,90]    [,91]    [,92]    [,93]    [,94]    [,95]    [,96]
#>   [1,] 235.5116 237.5630 239.6287 241.7056 243.7908 245.8813 247.9739 250.0655
#>           [,97]    [,98]    [,99]   [,100]   [,101]   [,102]   [,103]   [,104]
#>   [1,] 252.1530 254.2334 256.3037 258.3608 260.4018 262.4240 264.4247 266.4012
#>          [,105]   [,106]   [,107]   [,108]   [,109]   [,110]   [,111]   [,112]
#>   [1,] 268.3512 270.2724 272.1628 274.0204 275.8436 277.6310 279.3814 281.0938
#>          [,113]   [,114]   [,115]   [,116]   [,117]   [,118]   [,119]   [,120]
#>   [1,] 282.7676 284.4024 285.9982 287.5550 289.0736 290.5545 291.9991 293.4087
#>          [,121]   [,122]   [,123]   [,124]   [,125]   [,126]   [,127]   [,128]
#>   [1,] 294.7851 296.1303 297.4467 298.7370 300.0041 301.2512 302.4819 303.6997
#>          [,129]   [,130]   [,131]   [,132]   [,133]   [,134]   [,135]   [,136]
#>   [1,] 304.9087 306.1130 307.3170 308.5251 309.7421 310.9725 312.2213 313.4933
#>          [,137]   [,138]   [,139]   [,140]   [,141]   [,142]   [,143]   [,144]
#>   [1,] 314.7934 316.1264 317.4971 318.9104 320.3706 321.8824 323.4499 325.0772
#>          [,145]   [,146]   [,147]   [,148]   [,149]   [,150]   [,151]   [,152]
#>   [1,] 326.7680 328.5257 330.3535 332.2542 334.2299 336.2828 338.4143 340.6253
#>          [,153]   [,154]   [,155]   [,156]   [,157]   [,158]   [,159]   [,160]
#>   [1,] 342.9163 345.2873 347.7378 350.2665 352.8717 355.5513 358.3022 361.1210
#>          [,161]   [,162]   [,163]   [,164]   [,165]   [,166]   [,167]   [,168]
#>   [1,] 364.0037 366.9457 369.9417 372.9861 376.0725 379.1942 382.3439 385.5138
#>          [,169]   [,170]   [,171]   [,172]   [,173]   [,174]   [,175]   [,176]
#>   [1,] 388.6957 391.8811 395.0611 398.2264 401.3674 404.4744 407.5376 410.5468
#>          [,177]   [,178]   [,179]   [,180]   [,181]   [,182]   [,183]   [,184]
#>   [1,] 413.4919 416.3628 419.1494 421.8416 424.4295 426.9035 429.2540 431.4720
#>          [,185]   [,186]   [,187]   [,188]   [,189]   [,190]   [,191]   [,192]
#>   [1,] 433.5485 435.4753 437.2444 438.8483 440.2800 441.5333 442.6023 443.4820
#>          [,193]   [,194]   [,195]   [,196]   [,197]   [,198]   [,199]   [,200]
#>   [1,] 444.1678 444.6562 444.9440 445.0289 444.9093 444.5845 444.0543 443.3195
#>          [,201]   [,202]   [,203]   [,204]   [,205]   [,206]   [,207]   [,208]
#>   [1,] 442.3814 441.2422 439.9047 438.3726 436.6499 434.7417 432.6534 430.3910
#>          [,209]   [,210]   [,211]   [,212]   [,213]   [,214]   [,215]   [,216]
#>   [1,] 427.9612 425.3711 422.6283 419.7408 416.7170 413.5657 410.2958 406.9165
#>          [,217]   [,218]   [,219]   [,220]   [,221]   [,222]   [,223]   [,224]
#>   [1,] 403.4374 399.8679 396.2178 392.4967 388.7144 384.8804 381.0044 377.0958
#>          [,225]   [,226]   [,227]   [,228]   [,229]   [,230]   [,231]   [,232]
#>   [1,] 373.1637 369.2174 365.2655 361.3166 357.3789 353.4602 349.5681 345.7096
#>          [,233]   [,234]   [,235]   [,236]   [,237]   [,238]   [,239]   [,240]
#>   [1,] 341.8916 338.1202 334.4015 330.7407 327.1430 323.6128 320.1543 316.7710
#>          [,241]   [,242]   [,243]   [,244]   [,245]   [,246]   [,247]   [,248]
#>   [1,] 313.4662 310.2426 307.1026 304.0480 301.0804 298.2008 295.4099 292.7082
#>          [,249]   [,250]   [,251]   [,252]   [,253]    [,254]    [,255]
#>   [1,] 290.0955 287.5716 285.1359 282.7874 280.5250 278.34711 276.25219
#>           [,256]    [,257]    [,258]    [,259]    [,260]    [,261]    [,262]
#>   [1,] 274.23836 272.30355 270.44554 268.66199 266.95039 265.30817 263.73264
#>           [,263]    [,264]    [,265]    [,266]    [,267]    [,268]    [,269]
#>   [1,] 262.22105 260.77061 259.37846 258.04175 256.75760 255.52315 254.33554
#>           [,270]    [,271]    [,272]    [,273]    [,274]    [,275]    [,276]
#>   [1,] 253.19194 252.08958 251.02571 249.99765 249.00276 248.03850 247.10238
#>           [,277]    [,278]    [,279]    [,280]    [,281]    [,282]    [,283]
#>   [1,] 246.19200 245.30503 244.43925 243.59251 242.76275 241.94800 241.14639
#>           [,284]    [,285]    [,286]    [,287]    [,288]    [,289]    [,290]
#>   [1,] 240.35613 239.57554 238.80301 238.03703 237.27617 236.51910 235.76457
#>           [,291]    [,292]    [,293]    [,294]    [,295]    [,296]    [,297]
#>   [1,] 235.01140 234.25849 233.50483 232.74946 231.99153 231.23022 230.46479
#>           [,298]    [,299]    [,300]    [,301]    [,302]    [,303]    [,304]
#>   [1,] 229.69455 228.91889 228.13724 227.34909 226.55397 225.75146 224.94120
#>           [,305]    [,306]    [,307]    [,308]    [,309]    [,310]    [,311]
#>   [1,] 224.12285 223.29613 222.46076 221.61655 220.76329 219.90083 219.02903
#>           [,312]    [,313]    [,314]    [,315]    [,316]    [,317]    [,318]
#>   [1,] 218.14779 217.25703 216.35669 215.44673 214.52713 213.59789 212.65901
#>           [,319]    [,320]    [,321]    [,322]    [,323]    [,324]    [,325]
#>   [1,] 211.71054 210.75252 209.78499 208.80803 207.82171 206.82612 205.82136
#>           [,326]    [,327]    [,328]    [,329]    [,330]    [,331]    [,332]
#>   [1,] 204.80753 203.78476 202.75314 201.71282 200.66393 199.60660 198.54097
#>           [,333]    [,334]    [,335]    [,336]    [,337]    [,338]    [,339]
#>   [1,] 197.46720 196.38543 195.29582 194.19853 193.09372 191.98155 190.86220
#>           [,340]    [,341]    [,342]    [,343]    [,344]    [,345]    [,346]
#>   [1,] 189.73583 188.60261 187.46272 186.31634 185.16364 184.00481 182.84002
#>           [,347]    [,348]    [,349]    [,350]    [,351]    [,352]    [,353]
#>   [1,] 181.66946 180.49330 179.31174 178.12496 176.93314 175.73648 174.53516
#>           [,354]    [,355]    [,356]    [,357]    [,358]    [,359]    [,360]
#>   [1,] 173.32936 172.11928 170.90510 169.68702 168.46521 167.23988 166.01121
#>           [,361]    [,362]    [,363]    [,364]    [,365]    [,366]    [,367]
#>   [1,] 164.77940 163.54461 162.30706 161.06693 159.82440 158.57966 157.33290
#>           [,368]    [,369]    [,370]    [,371]    [,372]    [,373]    [,374]
#>   [1,] 156.08431 154.83407 153.58237 152.32940 151.07534 149.82036 148.56467
#>           [,375]    [,376]    [,377]    [,378]    [,379]    [,380]    [,381]
#>   [1,] 147.30843 146.05183 144.79505 143.53826 142.28165 141.02539 139.76966
#>           [,382]    [,383]    [,384]    [,385]    [,386]    [,387]    [,388]
#>   [1,] 138.51462 137.26046 136.00735 134.75545 133.50493 132.25596 131.00871
#>           [,389]    [,390]    [,391]    [,392]    [,393]    [,394]    [,395]
#>   [1,] 129.76334 128.52001 127.27888 126.04011 124.80387 123.57029 122.33955
#>           [,396]    [,397]    [,398]    [,399]    [,400]    [,401]    [,402]
#>   [1,] 121.11178 119.88715 118.66579 117.44785 116.23349 115.02283 113.81602
#>           [,403]    [,404]    [,405]    [,406]    [,407]    [,408]    [,409]
#>   [1,] 112.61321 111.41451 110.22008 109.03003 107.84450 106.66362 105.48751
#>           [,410]    [,411]    [,412]    [,413]    [,414]    [,415]   [,416]
#>   [1,] 104.31630 103.15010 101.98903 100.83322  99.68276  98.53779 97.39840
#>          [,417]   [,418]   [,419]   [,420]   [,421]   [,422]   [,423]   [,424]
#>   [1,] 96.26470 95.13679 94.01479 92.89878 91.78887 90.68515 89.58772 88.49666
#>          [,425]   [,426]   [,427]   [,428]   [,429]   [,430]   [,431]   [,432]
#>   [1,] 87.41206 86.33401 85.26259 84.19788 83.13997 82.08892 81.04481 80.00771
#>          [,433]   [,434]   [,435]   [,436]   [,437]   [,438]   [,439]   [,440]
#>   [1,] 78.97769 77.95482 76.93916 75.93077 74.92972 73.93605 72.94983 71.97110
#>          [,441]   [,442]   [,443]   [,444]   [,445]   [,446]   [,447]   [,448]
#>   [1,] 70.99992 70.03634 69.08039 68.13213 67.19158 66.25880 65.33382 64.41667
#>          [,449]   [,450]   [,451]   [,452]   [,453]    [,454]    [,455]
#>   [1,] 63.50739 62.60600 61.71253 60.82701 59.94946 59.079898 58.218349
#>           [,456]   [,457]    [,458]   [,459]    [,460]    [,461]    [,462]
#>   [1,] 57.364828 56.51935 55.681927 54.85257 54.031289 53.218091 52.412979
#>           [,463]    [,464]    [,465]    [,466]    [,467]    [,468]    [,469]
#>   [1,] 51.615959 50.827031 50.046193 49.273445 48.508782 47.752197 47.003683
#>           [,470]    [,471]    [,472]    [,473]    [,474]    [,475]    [,476]
#>   [1,] 46.263230 45.530828 44.806464 44.090122 43.381787 42.681441 41.989066
#>           [,477]    [,478]    [,479]    [,480]    [,481]    [,482]    [,483]
#>   [1,] 41.304639 40.628140 39.959544 39.298826 38.645959 38.000916 37.363668
#>           [,484]    [,485]    [,486]    [,487]    [,488]    [,489]    [,490]
#>   [1,] 36.734183 36.112430 35.498376 34.891987 34.293226 33.702057 33.118443
#>           [,491]    [,492]    [,493]    [,494]    [,495]    [,496]    [,497]
#>   [1,] 32.542344 31.973720 31.412530 30.858732 30.312284 29.773140 29.241256
#>           [,498]    [,499]    [,500]    [,501]
#>   [1,] 28.716586 28.199084 27.688701 27.185390
#>  [ reached getOption("max.print") -- omitted 149 rows ]

reconstructSp(Loadings2, Scores2, Spectra2)

#> hyperSpec object
#>    150 spectra
#>    3 data columns
#>    501 data points / spectrum
#> wavelength: lambda/nm [integer] 300 301 ... 800 
#> data:  (150 rows x 3 columns)
#>    1. gr:  [factor] B B ... A 
#>    2. class:  [factor] N l ... S1 
#>    3. spc: I / a.u. [matrix501] 160.1654 146.5726 ... 15.28009