[!] Simulate spectroscopic data

Simulation of spectroscopic data: Loadings (spectra of spectroscopic components), Scores (amplitudes of these components) an Spectra (mixture of the spectroscopic components multiplied by the amplitudes with random noise added).

simSpectra(
  x = 300:800,
  nGr = 2,
  nInGr = c(50, 50, 50),
  N = sum(nInGr),
  nDim = 4,
  w.possile = 5:150,
  plots = FALSE
)

Arguments

x
Values for x axis ("wavelengths").

nGr
Number of groups.

nInGr
Number of samples in a group (vector of values for each group).

N
Total number of samples.

nDim
Number of spectral components (dimensions).

w.possile
A vector of possible values of Gaussian curve parameter "w' (i.e., width). Values of w will be randomly sampled from this vector.

plots
Logical. If TRUE - makes plots. Default is FALSE.

Value

List of hyperSpec objects:

Spectra

Spectra of each observation, made of (Scores * Loadings) + NOISE.

Loadings

Normalized spectra of components.

Scores

Amplitudes of components for each observation.

In the list the hyperSpec objects contain spectroscopic data and additional variables.

Additional variables for Spectra and Scores:

class

A factor variable for classification.

gr

A factor variable for other type classification.

Additional variables for Loadings:

PeakAt

Position of components top peak.

cNames

Names of components.

order.of.rows

Original order of components before sorting.

Details

[This function works, but is not well documented yet.]

Examples

simSpectra()

#> $Spectra
#> hyperSpec object
#>    150 spectra
#>    3 data columns
#>    501 data points / spectrum
#> wavelength: list(lambda, nm) [integer] 300 301 ... 800 
#> data:  (150 rows x 3 columns)
#>    1. gr:  [factor] B B ... A 
#>    2. class:  [factor] K K ... S1 
#>    3. spc: I / a.u. [matrix501] 251.4165 190.9325 ... -5.083832 
#> 
#> $Loadings
#> hyperSpec object
#>    4 spectra
#>    4 data columns
#>    501 data points / spectrum
#> wavelength: list(lambda, nm) [integer] 300 301 ... 800 
#> data:  (4 rows x 4 columns)
#>    1. spc: I / a.u. [matrix501] 0.9915295 0.7027177 ... 0.005976023 
#>    2. cNames: cNames [character] max: 309nm max: 321nm max: 361nm max: 384nm 
#>    3. PeakAt: PeakAt [character] 309nm 321nm 361nm 384nm 
#>    4. order.of.rows: order.of.rows [integer] 1 2 3 4 
#> 
#> $Scores
#> hyperSpec object
#>    150 spectra
#>    3 data columns
#>    4 data points / spectrum
#> wavelength: Component [integer] 1 2 3 4 
#> data:  (150 rows x 3 columns)
#>    1. gr: gr [factor] B B ... A 
#>    2. class: class [factor] K K ... S1 
#>    3. spc: Amplitude [matrix4] 109.96971 65.83093 ... 48.99105 
#> 

simSpectra()$Spectra

#> hyperSpec object
#>    150 spectra
#>    3 data columns
#>    501 data points / spectrum
#> wavelength: list(lambda, nm) [integer] 300 301 ... 800 
#> data:  (150 rows x 3 columns)
#>    1. gr:  [factor] A A ... B 
#>    2. class:  [factor] K  S1 ... S1 
#>    3. spc: I / a.u. [matrix501] -0.6467433 -0.6117577 ... 243.1547