DEPRECATED!

DEPRECATED! Use cvo_test_bs instead.
Test if data in folds is stratified and blocked. This function run tests, that help to evaluate if data in folds is (a) stratified and (b) blocked.

test_folds_BS(
  obj,
  data = NULL,
  stratify_by = NULL,
  block_by = NULL,
  nColShow = 10
)

foldTests(obj, ...)

Arguments

obj
A list with validation/test set indices in folds. Note: If indices are from training set, the result will be incorrect.

data
A data frame, for wich obj was created.

stratify_by
A name of variable used for stratification.

block_by
A name of variable used for blocking.

nColShow
Number of blocking variable cross-tabulation's columns to be shown. Default is 10.

Value

Print tables that help to evaluate if data is (a) stratified, (b) blocked.

Examples

library(spHelper)

data(DataSet1)
obj <- createFoldsBS(data = DataSet1,
                     stratify_by = "gr", block_by = "ID",
                     returnTrain = FALSE)

#> Warning: 'createFoldsBS' is deprecated.
#> Use 'cvo_create_folds' instead.
#> See help("Deprecated")

cvo_test_bs(obj, DataSet1, stratify_by = "gr", block_by = "ID")

#> Error in cvo_test_bs(obj, DataSet1, stratify_by = "gr", block_by = "ID"): could not find function "cvo_test_bs"

#>  ************************************************************
#>      Test for STRATIFICATION
#>
#>        A B C D      <<<     >>>              A    B    C    D
#>  Fold1 2 2 2 2  <-Counts | Proportions->  0.25 0.25 0.25 0.25
#>  Fold2 2 2 2 2  <-Counts | Proportions->  0.25 0.25 0.25 0.25
#>  Fold3 2 2 2 2  <-Counts | Proportions->  0.25 0.25 0.25 0.25
#>  Fold4 2 2 2 2  <-Counts | Proportions->  0.25 0.25 0.25 0.25
#>  Fold5 2 2 2 2  <-Counts | Proportions->  0.25 0.25 0.25 0.25
#>
#>  If stratified, the proportions of each group in each fold
#>  (row) should be (approximately) equal and with no zero values.
#>  ____________________________________________________________
#>  Test for BLOCKING: BLOCKED
#>
#>        1 2 3 4 5 6 7 8 9 10 ..
#>  Fold1 0 0 0 0 2 0 0 2 0  0 ..
#>  Fold2 2 0 0 0 0 2 0 0 0  0 ..
#>  Fold3 0 0 2 0 0 0 0 0 2  0 ..
#>  Fold4 0 0 0 2 0 0 2 0 0  0 ..
#>  Fold5 0 2 0 0 0 0 0 0 0  2 ..
#>
#>  Number of unique IDs in each fold (first 10 columns).
#>  If blocked, the same ID appears just in one fold.
#>  ************************************************************